% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsg.R
\name{loadJsonSG}
\alias{loadJsonSG}
\title{Loads a json file rsg game specification
and returns the rsg object as R list
You then can call solveSG on the result
to get the approximations to the equilibrium payoff sets}
\usage{
loadJsonSG(file)
}
\arguments{
\item{file}{the file name of the json file}
}
\description{
Loads a json file rsg game specification
and returns the rsg object as R list
You then can call solveSG on the result
to get the approximations to the equilibrium payoff sets
}

