% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsg.R
\name{solveSG}
\alias{solveSG}
\title{Solve a two-player stochastic game of perfect monitoring}
\usage{
solveSG(delta = rsg$delta, states = rsg$states, rsg = NULL,
  duplicate.first.point = TRUE, all.iterations = FALSE, tol = 1e-12,
  normtol = tol, directiontol = tol, leveltol = tol, improvetol = tol)
}
\arguments{
\item{delta}{the discount factor between 0 and 1}

\item{stages}{a list of stages. Each stage is a list with the following elements:
- numActions: a size 2 vector that contains the number of actions for player 1 and 2. Its product is numActionProfiles.
- payoffs: a matrix of dimension numActionProfiles x 2. The payoffs as function of the action profile. The actions are mapped to action profiles as described in the documentation of Brooks C++ library: A pair (a1,a2) is mapped into an action profile index using the formula a=a1+a2*numActions[s][a1].
- transition: a matrix of dimension numActionProfiles x numStates. The transition probabilities. Each row has to sum up to 1.}
}
\value{
a list where points contains the extreme points
        of the final payoff set approximation for every state.
        Using the terminology of ABS (2016): the pivots from the last revolution.
        While SGSolve computes much more this R Interface only returns these values so far.
}
\description{
Uses Ben Brooks implementation SGSolve of the pencil sharpening
algorithm of Abreu, Brooks and Sanikov (2016)
See http://www.benjaminbrooks.net/sgsolvedoc/html/index.html
}

